submit
submit

+++ RenderBird/core/include/component_typelist.h
@@ -8,14 +8,14 @@ namespace Core
 	class ComponentTypeList : public Singleton<ComponentTypeList>
 	{
 	public:
-		using Iterator = std::map<int, TypeInfo*>::iterator;
+		using Iterator = std::map<size_t, TypeInfo*>::iterator;
 		ComponentTypeList();
 		void RegisterAllComponentType();
 		size_t GetComponentTypeCount()const;
-		TypeInfo* GetComponentTypeInfo(int typeId)const;
+		TypeInfo* GetComponentTypeInfo(size_t typeId)const;
 		Iterator begin() { return m_compTypeList.begin(); }
 		Iterator end() { return m_compTypeList.end(); }
 	private:
-		std::map<int, TypeInfo*> m_compTypeList;
+		std::map<size_t, TypeInfo*> m_compTypeList;
 	};
 }
